# generated/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Student": {
				"properties": {
					"studentName": {
						"type": "string"
					},
					"studentID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"studentName",
					"studentID"
				],
				"type": "object"
			},
			"Course": {
				"type": "string"
			},
			"CourseGrade": {
				"properties": {
					"grade": {
						"type": "number",
						"format": "double"
					},
					"course": {
						"$ref": "#/components/schemas/Course"
					}
				},
				"required": [
					"grade",
					"course"
				],
				"type": "object"
			},
			"Transcript": {
				"properties": {
					"grades": {
						"items": {
							"$ref": "#/components/schemas/CourseGrade"
						},
						"type": "array"
					},
					"student": {
						"$ref": "#/components/schemas/Student"
					}
				},
				"required": [
					"grades",
					"student"
				],
				"type": "object"
			},
			"InvalidParametersError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"db.StudentID": {
				"type": "number",
				"format": "double"
			},
			"db.CourseGrade": {
				"properties": {
					"grade": {
						"type": "number",
						"format": "double"
					},
					"course": {
						"$ref": "#/components/schemas/Course"
					}
				},
				"required": [
					"grade",
					"course"
				],
				"type": "object"
			},
			"db.Course": {
				"type": "string"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "module10-rest-transcript-activity",
		"version": "2023.2.0",
		"license": {
			"name": "BSD-3-Clause"
		},
		"contact": {
			"name": "Jonathan Bell"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/transcripts": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Transcript"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddStudent",
				"responses": {
					"200": {
						"description": "the ID of the newly created student",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"description": "Create a new student, optionally initializing their transcript",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"grades": {
										"items": {
											"$ref": "#/components/schemas/db.CourseGrade"
										},
										"type": "array"
									},
									"studentName": {
										"type": "string"
									}
								},
								"required": [
									"studentName"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/transcripts/{studentID}": {
			"get": {
				"operationId": "GetTranscript",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Transcript"
								}
							}
						}
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Retrieves the transcript for a given Student ID",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "studentID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/db.StudentID"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteStudent",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes a student's transcript",
				"security": [],
				"parameters": [
					{
						"description": "The ID of the student to delete",
						"in": "path",
						"name": "studentID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/db.StudentID"
						}
					}
				]
			}
		},
		"/transcripts/{studentID}/{courseNumber}": {
			"post": {
				"operationId": "AddGrade",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Adds a grade to a student's transcript",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "studentID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/db.StudentID"
						}
					},
					{
						"in": "path",
						"name": "courseNumber",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/db.Course"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"courseGrade": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"courseGrade"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetGrade",
				"responses": {
					"200": {
						"description": "the course grade",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"description": "Retrieve a student's grade in a course",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "studentID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/db.StudentID"
						}
					},
					{
						"in": "path",
						"name": "courseNumber",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/db.Course"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tags": [
		{
			"name": "transcript"
		}
	]
}